+++ .gitignore
@@ -13,3 +13,8 @@ node_modules/
 
 # Папка с собранными файлами проекта
 build/
+
+
+#bundle
+public/bundle.js
+public/bundle.js.map
