+++ src/components/card-generation.js
@@ -1,0 +1,10 @@
+import {getCardElement} from './task.js';
+
+
+export const generationCard = (count) => {
+  const cardArray = [];
+  for (let i = 0; i < count; i++) {
+    cardArray.push(getCardElement());
+  }
+  return cardArray.join(` `);
+};
