+++ src/components/load-more-button.js
@@ -1,0 +1,3 @@
+export const getLoadBtn = (text) => {
+  return `<button class="load-more" type="button">${text}</button>`;
+};
